vrrp: extract VIP actuation/reconcile to instance_vip.go
Pure code-motion split of pkg/vrrp/instance.go (#5661). Move the
netlink VIP add/remove actuation and the stale-VIP reconcile machinery
into a cohesive sibling file, package vrrp:

  vipRemoveReconcileMax, defaultVIPReconcileBackoff, surfaceStaleVIP,
  errReconcileSuperseded, removeVIPsIfBackup, scheduleVIPRemoveReconcile,
  vipActuationResult (+ ok), nlLinkByName, nlAddrAdd, nlAddrDel,
  addVIPsLocked, removeVIPs, removeVIPsLocked.

Declarations moved verbatim — no rename of exported symbols, no
logic/timing/const change, and no reorder of any goroutine, defer,
channel, or mutex operation (the reconcile goroutine, its generation
guard, and the *Locked helpers' lock expectations are unchanged).
Imports pruned by goimports; gofmt clean. Builds green. instance.go
drops from 2925 to 1392 LOC.
